kcqrs-client: few minor fixes in AggregateRepository and HttpEventStore
Performed few minor fixes in the AggregateRepository regarding error
handling and in the HttpEventStore.

+++ kcqrs-core/src/main/kotlin/com/clouway/kcqrs/core/EventCollisionException.kt
@@ -3,6 +3,6 @@ package com.clouway.kcqrs.core
 /**
  * @author Miroslav Genov (miroslav.genov@clouway.com)
  */
-class EventCollisionException(aggregateId: String, val version: Int, message: String) : AggregateException(aggregateId, message) {
-    constructor(aggregateId: String, version: Int) : this(aggregateId, version, "Data has been changed between loading and state changes.")
+class EventCollisionException(aggregateId: String, val version: Long, message: String) : AggregateException(aggregateId, message) {
+    constructor(aggregateId: String, version: Long) : this(aggregateId, version, "Data has been changed between loading and state changes.")
 }
